[ENGG-2443] refactor: common sidebar for api-client feature (#2326)
* refactor: api client context

* remove duplicate file

* fix: api client style

* cleanup

* refactor
