+++ app/src/features/apiClient/container.scss
@@ -1,0 +1,5 @@
+.api-client-container {
+  width: 100%;
+  height: 100%;
+  display: flex;
+}
