+++ app/src/features/apiClient/contexts/apiClient.tsx
@@ -3,8 +3,17 @@ import { useSelector } from "react-redux";
 import { getCurrentlyActiveWorkspace } from "store/features/teams/selectors";
 import { getUserAuthDetails } from "store/slices/global/user/selectors";
 import { RQAPI } from "../types";
-import { getApiRecords } from "backend/apiClient";
+import { getApiRecord, getApiRecords, upsertApiRecord } from "backend/apiClient";
 import Logger from "lib/logger";
+import { addToHistoryInStore, clearHistoryFromStore, getHistoryFromStore } from "../screens/apiClient/historyStore";
+import {
+  trackHistoryCleared,
+  trackImportCurlClicked,
+  trackNewRequestClicked,
+} from "modules/analytics/events/features/apiClient";
+import { useNavigate, useParams } from "react-router-dom";
+import { redirectToRequest } from "utils/RedirectionUtils";
+import { getEmptyAPIEntry } from "../screens/apiClient/utils";
 
 interface ApiClientContextInterface {
   apiClientRecords: RQAPI.Record[];
@@ -19,6 +28,21 @@ interface ApiClientContextInterface {
   isDeleteModalOpen: boolean;
   setIsDeleteModalOpen: React.Dispatch<React.SetStateAction<boolean>>;
   onDeleteModalClose: () => void;
+  history: RQAPI.Entry[];
+  addToHistory: (apiEntry: RQAPI.Entry) => void;
+  clearHistory: () => void;
+
+  isLoading: boolean;
+  selectedEntry: RQAPI.Entry;
+  selectedEntryDetails: RQAPI.ApiRecord;
+  isImportModalOpen: boolean;
+
+  onSelectionFromHistory: (index: number) => void;
+  saveRequest: (apiEntry: RQAPI.Entry) => Promise<void>;
+  handleImportRequest: (request: RQAPI.Request) => Promise<void>;
+  onImportClick: () => void;
+  onImportRequestModalClose: () => void;
+  onNewClick: (analyticEventSource: RQAPI.AnalyticsEventSource, recordType?: RQAPI.RecordType) => void;
 }
 
 const ApiClientContext = createContext<ApiClientContextInterface>({
@@ -34,6 +58,21 @@ const ApiClientContext = createContext<ApiClientContextInterface>({
   isDeleteModalOpen: false,
   setIsDeleteModalOpen: () => {},
   onDeleteModalClose: () => {},
+  history: [],
+  addToHistory: (apiEntry: RQAPI.Entry) => {},
+  clearHistory: () => {},
+
+  isLoading: false,
+  isImportModalOpen: false,
+  selectedEntry: undefined,
+  selectedEntryDetails: undefined,
+
+  onSelectionFromHistory: (index: number) => {},
+  saveRequest: async (apiEntry: RQAPI.Entry) => {},
+  handleImportRequest: async (request: RQAPI.Request) => {},
+  onImportClick: () => {},
+  onImportRequestModalClose: () => {},
+  onNewClick: (analyticEventSource: RQAPI.AnalyticsEventSource) => {},
 });
 
 interface ApiClientProviderProps {
@@ -41,6 +80,9 @@ interface ApiClientProviderProps {
 }
 
 export const ApiClientProvider: React.FC<ApiClientProviderProps> = ({ children }) => {
+  const { requestId } = useParams();
+  const navigate = useNavigate();
+
   const user = useSelector(getUserAuthDetails);
   const uid = user?.details?.profile?.uid;
   const workspace = useSelector(getCurrentlyActiveWorkspace);
@@ -49,6 +91,39 @@ export const ApiClientProvider: React.FC<ApiClientProviderProps> = ({ children }
   const [isLoadingApiClientRecords, setIsLoadingApiClientRecords] = useState(false);
   const [apiClientRecords, setApiClientRecords] = useState<RQAPI.Record[]>([]);
   const [recordToBeDeleted, setRecordToBeDeleted] = useState<RQAPI.Record>();
+  const [history, setHistory] = useState<RQAPI.Entry[]>(getHistoryFromStore());
+
+  const [isLoading, setIsLoading] = useState(false);
+  const [isImportModalOpen, setIsImportModalOpen] = useState(false);
+
+  const [selectedEntry, setSelectedEntry] = useState<RQAPI.Entry>();
+  const [selectedEntryDetails, setSelectedEntryDetails] = useState<RQAPI.ApiRecord>();
+
+  useEffect(() => {
+    if (!requestId || requestId === "new") {
+      return;
+    }
+
+    setSelectedEntry(null);
+    setIsLoading(true);
+
+    getApiRecord(requestId)
+      .then((result) => {
+        if (result.success) {
+          if (result.data.type === RQAPI.RecordType.API) {
+            setSelectedEntryDetails(result.data);
+          }
+        }
+      })
+      .catch((error) => {
+        setSelectedEntryDetails(null);
+        // TODO: redirect to new empty entry
+        Logger.error("Error loading api record", error);
+      })
+      .finally(() => {
+        setIsLoading(false);
+      });
+  }, [requestId]);
 
   useEffect(() => {
     if (!user.loggedIn) {
@@ -128,6 +203,77 @@ export const ApiClientProvider: React.FC<ApiClientProviderProps> = ({ children }
     setRecordToBeDeleted(null);
   }, []);
 
+  const addToHistory = useCallback((apiEntry: RQAPI.Entry) => {
+    setHistory((history) => [...history, apiEntry]);
+    addToHistoryInStore(apiEntry);
+  }, []);
+
+  const clearHistory = useCallback(() => {
+    setHistory([]);
+    clearHistoryFromStore();
+    trackHistoryCleared();
+  }, []);
+
+  const onSelectionFromHistory = useCallback(
+    (index: number) => {
+      setSelectedEntry(history[index]);
+    },
+    [history]
+  );
+
+  const saveRequest = useCallback(
+    async (apiEntry: RQAPI.Entry) => {
+      if (!user?.loggedIn) {
+        return;
+      }
+
+      setIsLoading(true);
+
+      const record: Partial<RQAPI.ApiRecord> = {
+        type: RQAPI.RecordType.API,
+        data: apiEntry,
+      };
+
+      const result = await upsertApiRecord(uid, record, teamId);
+
+      if (result.success) {
+        onSaveRecord(result.data);
+        redirectToRequest(navigate, result.data.id);
+      }
+
+      setIsLoading(false);
+    },
+    [uid, user?.loggedIn, teamId, onSaveRecord, navigate]
+  );
+
+  const handleImportRequest = useCallback(
+    async (request: RQAPI.Request) => {
+      const apiEntry = getEmptyAPIEntry(request);
+
+      return saveRequest(apiEntry)
+        .then(() => {
+          setSelectedEntry(apiEntry);
+        })
+        .finally(() => {
+          setIsImportModalOpen(false);
+        });
+    },
+    [saveRequest]
+  );
+
+  const onImportClick = useCallback(() => {
+    setIsImportModalOpen(true);
+    trackImportCurlClicked();
+  }, []);
+
+  const onImportRequestModalClose = useCallback(() => setIsImportModalOpen(false), []);
+
+  const onNewClick = useCallback((analyticEventSource: RQAPI.AnalyticsEventSource) => {
+    setSelectedEntry(getEmptyAPIEntry());
+    setSelectedEntryDetails(null);
+    trackNewRequestClicked(analyticEventSource);
+  }, []);
+
   const value = {
     apiClientRecords,
     isLoadingApiClientRecords,
@@ -141,6 +287,21 @@ export const ApiClientProvider: React.FC<ApiClientProviderProps> = ({ children }
     isDeleteModalOpen,
     setIsDeleteModalOpen,
     onDeleteModalClose,
+    history,
+    addToHistory,
+    clearHistory,
+
+    isLoading,
+    isImportModalOpen,
+    selectedEntry,
+    selectedEntryDetails,
+
+    onSelectionFromHistory,
+    saveRequest,
+    handleImportRequest,
+    onImportClick,
+    onImportRequestModalClose,
+    onNewClick,
   };
 
   return <ApiClientContext.Provider value={value}>{children}</ApiClientContext.Provider>;
