+++ app/src/features/apiClient/screens/apiClient/components/clientView/apiClientView.scss
@@ -7,12 +7,11 @@
   background-color: var(--requestly-color-background);
 
   .api-client-header-container {
-    padding: 20px 10px;
-    padding-bottom: 0;
+    padding: 8px 16px 0 6px;
   }
 
   .api-client-body {
-    padding: 0 10px;
+    padding: 0 16px;
     overflow: hidden;
   }
 
@@ -25,7 +24,6 @@
     display: flex;
     align-items: center;
     gap: 10px;
-    padding: 0 10px;
 
     .api-client-url-container {
       display: flex;
@@ -58,7 +56,6 @@
   }
 
   .api-request-tabs {
-    padding: 0 10px;
     height: calc(100% - 35px);
     overflow: hidden;
     margin-top: 8px;
