+++ app/src/features/apiClient/screens/apiClient/components/sidebar/components/collectionsList/collectionsList.scss
