+++ app/src/features/apiClient/screens/apiClient/components/sidebar/components/emptyState/emptyState.scss
