--- app/src/features/apiClient/screens/environment/components/environmentsSidebar/EnvironmentsSidebar.tsx
@@ -1,5 +1,0 @@
-import APIClientSidebar from "features/apiClient/screens/apiClient/components/sidebar/APIClientSidebar";
-
-export const EnvironmentsSidebar = () => {
-  return <APIClientSidebar />;
-};
