+++ app/src/features/apiClient/screens/environment/container.tsx
@@ -1,11 +1,9 @@
 import { Outlet } from "react-router-dom";
-import { EnvironmentsSidebar } from "./components/environmentsSidebar/EnvironmentsSidebar";
 import "./container.scss";
 
 export const EnvironmentContainer = () => {
   return (
     <div className="environment-container">
-      <EnvironmentsSidebar />
       <div className="environment-container-content">
         <Outlet />
       </div>
